+++ src/index.js
@@ -103,11 +103,6 @@ const createGlobalStateCommon = (initialState) => {
     if (process.env.NODE_ENV !== 'production') {
       validateName(name);
     }
-    const { ReactCurrentDispatcher } = __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED;
-    const dispatcher = ReactCurrentDispatcher.current;
-    if (dispatcher) {
-      throw new Error('getGlobalState should not be used in render. Consider useGlobalState.');
-    }
     return wholeGlobalState[name];
   };